refactoring QiscusEditText to view
